Fix softtabstop rendering
With soft tabstops, as much indent as possible is rendered with tabs, and the remainder with spaces. So, an indent of 5 x 4 = 20 is done via 2 * tab (16) + 4 * space, considering the 'tabstop' and 'softtabstop' values.

Add rainbow_levels#is_indented_with_softtabstop() and check that before rainbow_levels#is_indented_with_tabs() in rainbow_levels#get_pattern().