Correct issue with multiple indent levels that have the same color
Consider the following g:rainbow_levels:

let g:rainbow_levels = [
    \ {'ctermfg': 'red'},
    \ {'ctermfg': 'blue'},
    \ {'ctermfg': 'red'}]

Currently, this fails to apply red to both Level 1 and Level 3, since
get_highlight_command() uses index() to find the index of the color's
dictionary in g:rainbow_levels.  By replacing the for loop in
load_colors() with a while loop that keeps track of the index, the right
index for each indent level will be used, regardless of whether colors
are repeated or not.